Allow mocking instances instead of classes
Now given a Dog-class, you can either mock the class

  when(Dog).bark()...

so that all instances bark method become a stub, or a single instance

  rex = Dog()
  when(rex).bark()...

Then neither the class Dog nor a ?moritz=Dog()? are stubbed.

Implementation notes:
All Mock()s now point to a mocked_obj. In case of a Dummy, e.g. `mock = Mock()`, the mocked_obj is literally a mock.Dummy(). The so-called 'static_mocker' finally accepts all and everything - classes, modules, instances. Hence it is in fact a Registry for the known Mock()s, ensures that we only have one Mock-object per mocked object, and ensures that we can unstub() everything registered against it.
Hence the name 'static-mocker' is misleading and will change in refactorings.